--- src/modules/updateTodo.js
@@ -1,5 +1,0 @@
-function updateTodo(i, todos, newDesc) {
-  todos[i].desc = newDesc;
-  return todos;
-}
-module.exports = updateTodo;
