+++ src/modules/updateTodoList.js
@@ -1,0 +1,5 @@
+function updateTodo(i, todos, newDesc) {
+  todos[i].desc = newDesc;
+  return todos;
+}
+module.exports = updateTodo;
